Empty values in QueryParameterTransform and headers *Remove transforms (#978)
Adds support for empty string value in QueryParameterTransform. It changes the way the empty string "" value is handled, now it doesn't remove the given query parameter or header anymore, instead the "" is treated as a normal string value. To remove a query parameter the existing QueryParameterRemoveTransform can be used. To remove headers, the new transforms RequestHeaderRemoveTransform, ResponseHeaderRemoveTransform, and ResponseTrailerRemoveTransforms are added by this PR.
Note, setting "" as a header value is not recommended and can cause an undefined behavior.

Fixes #619